chore: extract DeferredCarWriter to go-car
Ref: https://github.com/ipld/go-car/pull/493
